merge from main

+++ test/e2e/ippoolcr/ippoolcr_test.go
@@ -31,12 +31,12 @@ var _ = Describe("test ippool CR", Label("ippoolCR"), func() {
 	BeforeEach(func() {
 		if frame.Info.SpiderSubnetEnabled {
 			if frame.Info.IpV4Enabled {
-				v4SubnetName, v4SubnetObject = common.GenerateExampleV4SubnetObject(20)
+				v4SubnetName, v4SubnetObject = common.GenerateExampleV4SubnetObject(frame, 20)
 				Expect(v4SubnetObject).NotTo(BeNil())
 				Expect(common.CreateSubnet(frame, v4SubnetObject)).NotTo(HaveOccurred())
 			}
 			if frame.Info.IpV6Enabled {
-				v6SubnetName, v6SubnetObject = common.GenerateExampleV6SubnetObject(20)
+				v6SubnetName, v6SubnetObject = common.GenerateExampleV6SubnetObject(frame, 20)
 				Expect(v6SubnetObject).NotTo(BeNil())
 				Expect(common.CreateSubnet(frame, v6SubnetObject)).NotTo(HaveOccurred())
 			}
@@ -79,6 +79,11 @@ var _ = Describe("test ippool CR", Label("ippoolCR"), func() {
 		}
 
 		DeferCleanup(func() {
+			if CurrentSpecReport().Failed() {
+				GinkgoWriter.Println("If the use case fails, the cleanup step will be skipped")
+				return
+			}
+
 			GinkgoWriter.Println("Clean test ENV")
 			err = frame.DeleteNamespace(nsName)
 			Expect(err).NotTo(HaveOccurred(), "Failed to delete namespace %v, err: %v", nsName, err)
