+++ test/e2e/kruise/kruise_test.go
@@ -41,18 +41,23 @@ var _ = Describe("Third party control: OpenKruise", Label("kruise"), func() {
 
 		if frame.Info.SpiderSubnetEnabled {
 			if frame.Info.IpV4Enabled {
-				v4SubnetName, v4SubnetObject = common.GenerateExampleV4SubnetObject(IpNum)
+				v4SubnetName, v4SubnetObject = common.GenerateExampleV4SubnetObject(frame, IpNum)
 				Expect(v4SubnetObject).NotTo(BeNil())
 				Expect(common.CreateSubnet(frame, v4SubnetObject)).NotTo(HaveOccurred())
 			}
 			if frame.Info.IpV6Enabled {
-				v6SubnetName, v6SubnetObject = common.GenerateExampleV6SubnetObject(IpNum)
+				v6SubnetName, v6SubnetObject = common.GenerateExampleV6SubnetObject(frame, IpNum)
 				Expect(v6SubnetObject).NotTo(BeNil())
 				Expect(common.CreateSubnet(frame, v6SubnetObject)).NotTo(HaveOccurred())
 			}
 		}
 
 		DeferCleanup(func() {
+			if CurrentSpecReport().Failed() {
+				GinkgoWriter.Println("If the use case fails, the cleanup step will be skipped")
+				return
+			}
+
 			GinkgoWriter.Printf("delete namespace %v. \n", namespace)
 			Expect(frame.DeleteNamespace(namespace)).NotTo(HaveOccurred())
